Refactoring, move old DynmapServer into LocalDynmapServer and create an interface for DynmapServer to allow using remote dynmap instances later.
